revert: the vague-linkage demotion; COMDAT proves the library, not consumers
`extern template` breaks the invariant, and it is ordinary code rather
than an edge case. A public header carrying

    extern template struct Box<int>;

tells every consumer TU *not* to instantiate, while the library's own
explicit instantiation still emits a weak COMDAT definition. Verified
against g++:

    library  object: _ZNK3BoxIiE3getEv  in a COMDAT group
    consumer object: NOTYPE GLOBAL UND, COMDAT set empty

So dropping that export breaks the consumer, and the predicate demoted it
to a risk (Codex review).

The demotion's invariant is "every consumer already emitted its own
copy". COMDAT membership establishes that the *library* used vague
linkage -- a different statement. This is the same shape as the first
attempt, which read `is_inline` (a specifier) as proof of a definition:
both times a fact about the library's own build was read as a fact about
its consumers. Nothing in two library snapshots, and nothing in one
library's object files, separates "the consumer emitted a copy" from "the
consumer holds an undefined reference".

Kept: `buildsource/comdat_groups.py` and its `BuildEvidence.comdat`
collection. The parser is sound, byte-order correct, ELF-only by
construction, and answers a real question -- just not this one.
`comdat_group` was already reserved vocabulary in `graph_facts.py` for a
future linker-artifact extractor; this is that parser.

AGENTS.md records both counterexamples with their empirical checks, the
shape they share, and what a third attempt would actually need
(consumer-side evidence, or an `extern template` header fact castxml does
not expose).
