+++ abicheck/diff_symbols.py
@@ -109,7 +109,6 @@ from .model import (
     stdlib_namespaces_excluded,
 )
 from .name_classification import is_local_rtti_symbol
-from .symbol_linkage import check_removed_function, vague_linkage_symbols
 
 # Visibility levels that constitute the public ABI surface.
 _PUBLIC_VIS = (Visibility.PUBLIC, Visibility.ELF_ONLY)
@@ -263,6 +262,39 @@ def _format_params(params: list[Param]) -> str:
     """
     parts = [p.type for p in params]
     return ", ".join(parts) if parts else "(none)"
+
+
+def _check_removed_function(
+    mangled: str,
+    f_old: Function,
+    new_all: dict[str, Function],
+    elf_only_mode: bool,
+) -> Change:
+    """Create a Change for a function that was removed or hidden."""
+    f_hidden = new_all.get(mangled)
+    if (
+        f_hidden is not None
+        and f_hidden.visibility == Visibility.HIDDEN
+        and not (elf_only_mode and f_old.visibility == Visibility.ELF_ONLY)
+    ):
+        return make_change(
+            ChangeKind.FUNC_VISIBILITY_CHANGED,
+            symbol=mangled,
+            name=f_old.name,
+            old_value=f_old.visibility.value,
+            new_value=f_hidden.visibility.value,
+        )
+    removed_kind = (
+        ChangeKind.FUNC_REMOVED_ELF_ONLY
+        if (elf_only_mode and f_old.visibility == Visibility.ELF_ONLY)
+        else ChangeKind.FUNC_REMOVED
+    )
+    return make_change(
+        removed_kind,
+        symbol=mangled,
+        description=f"{f_old.visibility.value.capitalize()} function removed: {f_old.name}",
+        old_value=f_old.name,
+    )
 
 
 def _check_return_type_change(
@@ -723,7 +755,6 @@ def _match_old_function(
     elf_only_mode: bool,
     params_unconfirmed: bool = False,
     is_llp64: bool = False,
-    vague_symbols: frozenset[str] | None = None,
 ) -> list[Change]:
     """Classify a single old function: matched by mangled, extern-C fallback, or removed.
 
@@ -789,11 +820,7 @@ def _match_old_function(
         matched_by_name.add(f_old.name)
         return result
 
-    return [
-        check_removed_function(
-            mangled, f_old, new_all, elf_only_mode, vague_symbols
-        )
-    ]
+    return [_check_removed_function(mangled, f_old, new_all, elf_only_mode)]
 
 
 def _detect_newly_deleted_functions(
@@ -900,10 +927,6 @@ def _diff_functions(old: AbiSnapshot, new: AbiSnapshot) -> list[Change]:
     new_all = new.function_map
 
     matched_by_name: set[str] = set()
-    # Proven vague-linkage symbols from the *old* side: the consumer was built
-    # against the old headers, so only that build can establish it emitted its
-    # own COMDAT copy. None when there is no L3 evidence -- never a demotion.
-    vague_symbols = vague_linkage_symbols(old)
 
     for mangled, f_old in old_map.items():
         changes.extend(
@@ -916,7 +939,6 @@ def _diff_functions(old: AbiSnapshot, new: AbiSnapshot) -> list[Change]:
                 elf_only_mode,
                 params_unconfirmed,
                 is_llp64,
-                vague_symbols,
             )
         )
