+++ abicheck/reporter_markdown.py
@@ -150,9 +150,6 @@ _OPERATION_OVERRIDES: dict[str, str] = {
     # A whole ISA-dispatch family's concrete symbols vanish (case83), not a
     # property change on a persisting symbol.
     "cpu_dispatch_isa_dropped": "removed",
-    # Ends in "_dropped" but the suffix list keys on "_removed"; the event is
-    # still an export disappearing, only at a lower severity.
-    "func_vague_export_dropped": "removed",
     # A stable name is added alongside the still-present experimental alias
     # (case99) -- without the dedicated detector this would just be a plain
     # func_added; ADDITION_KINDS already classifies it as an addition
