+++ docs/contribute/abicc-test-coverage-comparison.md
@@ -1,6 +1,6 @@
 # ABICC vs Abicheck: Test Coverage Comparison
 
-> Updated: 2026-03-09; content reviewed 2026-06-07. The ABICC rule mapping is a historical parity snapshot; the current ChangeKind total is **397**, see the [Change Kind Reference](../reference/change-kinds.md).
+> Updated: 2026-03-09; content reviewed 2026-06-07. The ABICC rule mapping is a historical parity snapshot; the current ChangeKind total is **396**, see the [Change Kind Reference](../reference/change-kinds.md).
 > Source: ABICC `RulesBin.xml` (196 rules), `RulesSrc.xml` (100 rules + `Removed_Const_Overload`), `RegTests.pm` (~153 C++ + ~102 C named scenarios)
 > Target: abicheck's three coverage surfaces —
 > `examples/` (195 ground-truth entries today: 159 binary shared-library
@@ -8,7 +8,7 @@
 > fixture/source-only L2/L5/source cases; the original 74-case subset is the
 > release-pinned cross-tool benchmark — see [Tool Comparison](../reference/tool-comparison.md)
 > for how the split is derived and kept in sync), `tests/` (large
-> unit/integration/parity suite), and the `ChangeKind` enum (**397 change
+> unit/integration/parity suite), and the `ChangeKind` enum (**396 change
 > kinds** today; the per-rule mappings below were written against an earlier
 > snapshot)
 >
@@ -30,7 +30,7 @@
 | ABICC RegTests.pm named scenarios | ~255 (~153 C++ + ~102 C) |
 | ABICC de-duplicated scenarios | ~66 |
 | **Abicheck covers (has ChangeKind + tests)** | **66/66 (100%)** |
-| Abicheck ChangeKind enum members | **397 change kinds** (this table's per-rule mappings reflect an earlier snapshot) |
+| Abicheck ChangeKind enum members | **396 change kinds** (this table's per-rule mappings reflect an earlier snapshot) |
 | All ChangeKinds have assertion tests | **Yes** (enforced by `test_changekind_completeness.py`) |
 | Abicheck example cases | 193 total: 159 binary shared-library competitor lanes + 34 dedicated non-.so lanes (including fixture/source-only L2/L5/source cases) |
 | ABICC scenarios NOT in abicheck | **0** |
@@ -381,6 +381,6 @@ These detectors exist in abicheck but have no ABICC equivalent:
 
 **0 remaining gaps.** All ABICC de-duplicated detection scenarios are covered by abicheck with dedicated ChangeKinds and explicit assertion tests, including the `TypedefToFunction` scenario (covered in `test_changekind_completeness.py`).
 
-**All 397 `ChangeKind`s have assertion-level test coverage today**, enforced by `test_changekind_completeness.py`. Previously, 3 ChangeKinds (`SYMBOL_BINDING_STRENGTHENED`, `VAR_ACCESS_WIDENED`, `TYPE_VTABLE_CHANGED`) were only referenced in set/list definitions but lacked explicit assertion tests; these are now covered.
+**All 396 `ChangeKind`s have assertion-level test coverage today**, enforced by `test_changekind_completeness.py`. Previously, 3 ChangeKinds (`SYMBOL_BINDING_STRENGTHENED`, `VAR_ACCESS_WIDENED`, `TYPE_VTABLE_CHANGED`) were only referenced in set/list definitions but lacked explicit assertion tests; these are now covered.
 
 **Inline function scenarios (4):** `RemovedInlineMethod`, `removedInlineFunction`, `functionBecameInline`, `RemovedInlineVirtualFunction` — these ABICC scenarios detect inline function removal via header comparison. In abicheck, inline functions declared in headers are parsed by castxml but filtered against ELF `.dynsym` (inline functions have no exported symbol). If headers are provided, castxml captures the declaration; detection depends on whether the symbol was previously exported. Virtual inline function removal is still detected via vtable changes (`TYPE_VTABLE_CHANGED`). These are classified as **edge cases** rather than gaps, since the typical ABI contract concerns exported symbols.
