+++ docs/reference/detector-spec.json
@@ -1,6 +1,6 @@
 {
   "description": "Formal detector specification matrix (generated).",
-  "count": 397,
+  "count": 396,
   "detectors": [
     {
       "kind": "abi_relevant_build_flag_changed",
@@ -1153,15 +1153,6 @@
       "min_evidence": "L1",
       "doc_slug": "func_static_changed",
       "examples": "case21_method_became_static"
-    },
-    {
-      "kind": "func_vague_export_dropped",
-      "category": "risk",
-      "default_verdict": "COMPATIBLE_WITH_RISK",
-      "severity": "warning",
-      "min_evidence": "L3",
-      "doc_slug": "func_vague_export_dropped",
-      "examples": ""
     },
     {
       "kind": "func_variadic_added",
