+++ tests/test_changekind_completeness.py
@@ -90,8 +90,6 @@ ASSERTED_CHANGE_KINDS: set[ChangeKind] = {
     ChangeKind.FUNC_PURE_VIRTUAL_ADDED,
     ChangeKind.FUNC_REMOVED,
     ChangeKind.FUNC_REMOVED_ELF_ONLY,
-    # Covered by tests/test_vague_linkage_export.py
-    ChangeKind.FUNC_VAGUE_EXPORT_DROPPED,
     ChangeKind.FUNC_RETURN_CHANGED,
     ChangeKind.FUNC_STATIC_CHANGED,
     ChangeKind.FUNC_VIRTUAL_ADDED,
